doc/css: adjust code and userinput styles
I don't know why <code> was defined to be bold, it looks bad to me. The
same with <synopsis> which would inherit bold from <dt> when used inside
a variablelist term. So, to make the code snippets look better, force
them to use normal weight.

<userinput> should differentiate from normal code somehow, and italic
seems fine for it.

<literal> already has bold, and I think it's fine, so no need to touch
it.

These changes are mainly for the new XML dialect documentation chapter.
I didn't notice anything changing for the older or generated parts of the
docs.

Signed-off-by: Pekka Paalanen <pq@iki.fi>